requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
tetpyclient==1.0.8
urllib3==2.1.0
